refactor(cli/runner): move runner Context class to cli and rename to CLIContext
